Add clearCompletedTasks functionality to 'Clear all' button
