+++ src/style.css
@@ -108,16 +108,16 @@ li#list-title {
   color: blue;
 }
 
-.completed .description {
-  text-decoration: line-through;
-  color: grey;
-}
-
 .description {
   font-family: Arial, Helvetica, sans-serif;
   width: 100%;
   height: 100%;
   border: none;
+}
+
+.completed .description {
+  text-decoration: line-through;
+  color: grey;
 }
 
 .description:focus-visible {
